Crate gdkwayland

Rust GDK3-Wayland bindings

Project site is here.

Rust bindings and wrappers for GDK3-Wayland, part of gtk3-rs.

GDK3-Wayland 3.22 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.92.0.

Documentation

Rust API - Stable

Rust API - Development

C API

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
gdkwayland = { git = "https://github.com/gtk-rs/gtk3-rs.git", package = "gdkwayland" }

Avoid mixing versioned and git crates like this:

# This will not compile
[dependencies]
gdkwayland = "0.15"
gdkwayland = { git = "https://github.com/gtk-rs/gtk3-rs.git", package = "gdkwayland" }

See Also

glib

cairo-rs

gdk-pixbuf

gio

pango

License

gdkwayland is available under the MIT License, please refer to it.

Re-exports

pub use gdk;

pub use gdk_wayland_sys as ffi;

Structs

WaylandDevice

GLib type: GObject with reference counted clone semantics.

WaylandDisplay

GLib type: GObject with reference counted clone semantics.

WaylandGLContext

GLib type: GObject with reference counted clone semantics.

WaylandMonitor

GLib type: GObject with reference counted clone semantics.

WaylandSeat

GLib type: GObject with reference counted clone semantics.

WaylandWindow

GLib type: GObject with reference counted clone semantics.