Crate gdkwayland

Rust GDK3-Wayland bindings

Project site is here.

Rust bindings and wrappers for GDK3-Wayland, part of gtk3-rs.

GDK3-Wayland 3.18 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.56.0.

Documentation

Rust API - Stable

Rust API - Development

C API

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
gdkwayland = { git = "https://github.com/gtk-rs/gtk3-rs.git", package = "gdkwayland" }

Avoid mixing versioned and git crates like this:

[dependencies]
gdkwayland = "0.15"
gdkwayland = { git = "https://github.com/gtk-rs/gtk3-rs.git", package = "gdkwayland" }

See Also

glib

cairo-rs

gdk-pixbuf

gio

pango

License

gdk_waylang is available under the MIT License, please refer to it.

Re-exports

pub use ffi;

pub use gdk;

Structs

WaylandDevice

WaylandDisplay

WaylandGLContext

WaylandMonitor

WaylandSeat

WaylandWindow